refactor: separate other concerns from ServerRenderProductController
- ServerRenderProductController에서 ExceptionHandler 로직을 별도로 분리
- root 페이지 접속시 products로 리다이렉트하는 기능을 별도의 컨트롤러로 분리
