+++ src/main/java/gift/global/GlobalExceptionHandler.java
@@ -3,19 +3,19 @@ package gift.global;
 import gift.domain.product.exception.ProductAlreadyExistsException;
 import gift.domain.product.exception.ProductNotFoundException;
 import gift.global.response.ErrorResponse;
-import jakarta.validation.constraints.NotNull;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 /**
  * 서버에서 발생하는 예외를 종합적으로 처리하는 클래스
  */
-@RestControllerAdvice
+@RestControllerAdvice(annotations = RestController.class)
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(ProductNotFoundException.class)
