+++ src/main/java/gift/server/rendering/controller/ServerRenderEntryRedirectController.java
@@ -1,0 +1,13 @@
+package gift.server.rendering.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ServerRenderEntryRedirectController {
+
+    @GetMapping("/")
+    public String index() {
+        return "redirect:/products";
+    }
+}
