+++ src/main/java/gift/server/rendering/controller/ServerRenderProductController.java
@@ -1,53 +1,51 @@
-package gift.serverside.rendering;
+package gift.server.rendering.controller;
 
 import gift.domain.product.ProductService;
 import gift.domain.product.dto.ProductRequestDto;
 import gift.domain.product.dto.ProductResponseDto;
-import gift.domain.product.exception.ProductAlreadyExistsException;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-public class ProductController {
+@RequestMapping("/products")
+public class ServerRenderProductController {
 
     private final ProductService service;
 
     @Autowired
-    public ProductController(ProductService service) {
+    public ServerRenderProductController(ProductService service) {
         this.service = service;
     }
 
-    @GetMapping("/")
-    public String index() {
-        return "redirect:/products";
-    }
-
-    @GetMapping("/products")
+    @GetMapping
     public String showProducts(Model model) {
         List<ProductResponseDto> products = service.getAllProducts();
         model.addAttribute("products", products);
         return "products";
     }
 
-    @GetMapping("/products/add")
+    @GetMapping("/add")
     public String showAddProductForm(Model model) {
         model.addAttribute("productRequestDto", new ProductRequestDto("", 0L, ""));
         return "addProduct";
     }
 
-    @PostMapping("/products/add")
+    @PostMapping("/add")
     public String addProduct(@Valid @ModelAttribute ProductRequestDto requestDto, Model model) {
         service.addProduct(requestDto);
         return "redirect:/products";
     }
 
-    @GetMapping("/products/update/{id}")
+    @GetMapping("/update/{id}")
     public String showUpdateProductForm(@PathVariable("id") Long id, Model model) {
         ProductResponseDto product = service.getProductById(id);
         ProductRequestDto dto = new ProductRequestDto(product.name(), product.price(), product.imageUrl());
@@ -56,35 +54,15 @@ public class ProductController {
         return "updateProduct";
     }
 
-    @PutMapping("/products/update/{id}")
+    @PutMapping("/update/{id}")
     public String updateProduct(@PathVariable("id") Long id, @Valid @ModelAttribute ProductRequestDto requestDto) {
         service.updateProductById(id, requestDto);
         return "redirect:/products";
     }
 
-    @GetMapping("/products/delete/{id}")
+    @GetMapping("/delete/{id}")
     public String deleteProduct(@PathVariable("id") Long id) {
         service.deleteProduct(id);
         return "redirect:/products";
     }
-
-    //이 컨트롤러만을 위한 non-RESTful한 익셉션 핸들러
-    @ResponseStatus(HttpStatus.CONFLICT)
-    @ExceptionHandler(ProductAlreadyExistsException.class)
-    public String handleProductAlreadyExistsException(ProductAlreadyExistsException e, Model model) {
-        model.addAttribute("headTitle", "상품 중복");
-        model.addAttribute("pageTitle", "상품 중복 오류");
-        model.addAttribute("errorMessage", "이름, 가격, 이미지 URL이 같은 상품이 이미 존재합니다. 중복이 아닌 상품을 입력해주세요.");
-        return "errorDisplayPage";
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public String handleMethodArgumentNotValidException(MethodArgumentNotValidException e, Model model) {
-        model.addAttribute("headTitle", "입력 오류");
-        model.addAttribute("pageTitle", "입력 오류");
-        model.addAttribute("errorMessage", e.getBindingResult().getFieldError().getDefaultMessage());
-        return "errorDisplayPage";
-    }
-
 }
